+++ hooks/usePageView.tsx
@@ -1,0 +1,23 @@
+import { useRouter } from "next/router";
+import { useEffect } from "react";
+import { existsGaId, pageview } from "../lib/gtag";
+
+export default function usePageView() {
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!existsGaId) {
+      return;
+    }
+
+    const handleRouteChange = (path) => {
+      pageview(path);
+    };
+
+    router.events.on("routeChangeComplete", handleRouteChange);
+
+    return () => {
+      router.events.off("routeChangeComplete", handleRouteChange);
+    };
+  }, [router.events]);
+}
